requirements.txt
google-cloud-aiplatform==1.71.1
google-genai==1.72.0
langchain_chroma==1.1.0
langchain_community==0.4.1
langchain_google_genai==4.2.1